attempting to brute force display of values
So I really need a display meter for some demos, which is a larger task than I have time for right now.

This doesn't work, will need to find a different react-friendly approach but no harm in pushing this to main for the time being.

I will be content to include a non-react display somwhere in the patcher so that I can show people the midi signals coming in.